Fix incorrect left join in database queries
Turns out the 'mapid' field refers to the 'map_version' table, not
'map'. While fixing that, we clean up the database setup code a tiny
bit and ensure that the issue will be fixed in the future.

TODO: find a way to express our test data in a less messy and confusing
way.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>