Remove ConnectionReplayStream
We don't need a subclass for this - apart from reading data from a
connection it behaves exactly the same as a stream fed from outside.
Replace it with a Reader class that reads from the connection into a
stream.

Also add a convenience function to merge strategies that does all the
strategy callbacks for us when given a replay stream. It's so simple
that unit tests are probably useless for it.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>