Don't consider errors for connections we closed
If we closed a connection on our end, then malformed (e.g. short)
connections shouldn't really be considered errors and we shouldn't treat
them as such.

FIXME: a unit test was modified for coverage and nothing else since
logging and metrics are singletons with no side effects. Is that ugly?

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>